refactor: Moved model train/eval scripts to new subpackages
- Add SimulationCache class to remove global keyword from web.py
